python/tmp/plone-training/conference-site/backend/src/conference_site/src/conference_site/__init__.py
"""Init and utils."""
from zope.i18nmessageid import MessageFactory

import logging


PACKAGE_NAME = "conference_site"

_ = MessageFactory("conference_site")

logger = logging.getLogger("conference_site")
